==> train/notebooks/notebook_04.ipynb <==
import os
import pandas as pd

from surprise import Dataset, Reader
from surprise.model_selection import train_test_split, cross_validate

file_path_metadata = "../data/news-portal-user-interactions-by-globocom/articles_metadata.csv"
df_metadata = pd.read_csv(file_path_metadata)
df_metadata = pd.DataFrame(df_metadata)
df_metadata

file_path_clicks = "../data/news-portal-user-interactions-by-globocom/clicks_sample.csv"
df_click_sample = pd.read_csv(file_path_clicks)
df = pd.DataFrame(df_click_sample)
df_click_sample

clicks_directory = "../data/news-portal-user-interactions-by-globocom/clicks/"
columns = ["user_id", "session_id", "session_start", "session_size", "click_article_id",
           "click_timestamp", "click_environment", "click_deviceGroup", "click_os",
           "click_country", "click_region", "click_referrer_type"]
df_clicks = pd.DataFrame(columns = columns)

for i in os.listdir(clicks_directory):
    temp_ = pd.read_csv(clicks_directory + i, index_col=None, header=0)
    liste_concat = [df_clicks, temp_]
    df_clicks = pd.concat(liste_concat, ignore_index=True)


df_clicks

df_merged = df_clicks.merge(df_metadata, left_on='click_article_id', right_on='article_id')
df_merged = df_merged[['user_id', 'article_id', 'category_id']]
df_merged["value"] = 1
df_merged

grouped = df_merged.groupby(['user_id', 'category_id'])
df_aggregated = grouped['value'].sum().reset_index()
df_aggregated = df_aggregated.rename(columns={'value': 'rate'})
df_aggregated

df_aggregated.rate.unique()

final_df = '../data/csv/df_aggregated.csv'
df_aggregated.to_csv(final_df, index=False)

df_aggregated = pd.read_csv('../data/csv/df_aggregated.csv')

print("Nombre de valeurs pour 1:", len(df_aggregated[df_aggregated["rate"] == 1]))
print("Nombre de valeurs pour 2:", len(df_aggregated[df_aggregated["rate"] == 2]))
print("Nombre de valeurs pour 3:", len(df_aggregated[df_aggregated["rate"] == 3]))
print("Nombre de valeurs pour 4:", len(df_aggregated[df_aggregated["rate"] == 4]))
print("Nombre de valeurs pour 5:", len(df_aggregated[df_aggregated["rate"] == 5]))
print("Nombre de valeurs supérieur à 5:", len(df_aggregated[df_aggregated["rate"] > 5]))

df_aggregated.loc[df_aggregated["rate"] == 1, "rate"] = 0
df_aggregated.loc[df_aggregated["rate"] == 2, "rate"] = 1
df_aggregated.loc[df_aggregated["rate"] == 3, "rate"] = 2
df_aggregated.loc[df_aggregated["rate"] == 4, "rate"] = 3
df_aggregated.loc[df_aggregated["rate"] == 5, "rate"] = 4
df_aggregated.loc[df_aggregated["rate"] > 5, "rate"] = 5

print("Nombre de valeurs pour 0:", len(df_aggregated[df_aggregated["rate"] == 0]))
print("Nombre de valeurs pour 1:", len(df_aggregated[df_aggregated["rate"] == 1]))
print("Nombre de valeurs pour 2:", len(df_aggregated[df_aggregated["rate"] == 2]))
print("Nombre de valeurs pour 3:", len(df_aggregated[df_aggregated["rate"] == 3]))
print("Nombre de valeurs pour 4:", len(df_aggregated[df_aggregated["rate"] == 4]))
print("Nombre de valeurs pour 5:", len(df_aggregated[df_aggregated["rate"] == 5]))

df_sample = pd.DataFrame(columns=["user_id", "category_id", "rate"])

for i in range(6):
    df_sample = pd.concat([df_sample, df_aggregated[df_aggregated["rate"] == i].sample(20000)],
                          ignore_index=True)

reader = Reader(rating_scale=(1,5))
df_significant = df_sample.copy()
data = Dataset.load_from_df(df_significant[['user_id', 'category_id', 'rate']], reader)

print('We have selects', len(df_significant), 'interactions.')

trainset, testset = train_test_split(data, test_size=0.25)
print('Test set lenght :', len(testset))
print('Train set lenght :', len(df_significant) - len(testset))

from surprise import SVD, accuracy

algo = SVD()
algo.fit(trainset)

predictions = algo.test(testset)
print('Number of predictions in Test set :', len(predictions))

accuracy.rmse(predictions)

df_predict = []
for i in range(len(predictions)):
    df_predict.append([
        predictions[i].uid,
        predictions[i].iid,
        predictions[i].r_ui,
        predictions[i].est]
    )

df_predict = pd.DataFrame(df_predict, columns=["user_id", "category_id", "y_true", "y_predict"])
df_predict

temporary_ = df_predict[df_predict["category_id"] == 375].copy()
temporary_

temporary_["y_predict"].max()

temporary_[temporary_["y_predict"]==4.800513885620038]
