==> train/notebooks/notebook_09.ipynb <==
import pandas as pd
import numpy as np
import os

pickle = pd.read_pickle('../data/news-portal-user-interactions-by-globocom/articles_embeddings.pickle')
pickle = pd.DataFrame(pickle)
pickle

clicks_directory = "../data/news-portal-user-interactions-by-globocom/clicks/"
columns = ["user_id", "session_id", "session_start", "session_size", "click_article_id",
           "click_timestamp", "click_environment", "click_deviceGroup", "click_os",
           "click_country", "click_region", "click_referrer_type"]
df_clicks = pd.DataFrame(columns = columns)

for i in os.listdir(clicks_directory):
    temp_ = pd.read_csv(clicks_directory + i, index_col=None, header=0)
    liste_concat = [df_clicks, temp_]
    df_clicks = pd.concat(liste_concat, ignore_index=True)
df_clicks

file_path_metadata = "../data/news-portal-user-interactions-by-globocom/articles_metadata.csv"
df_metadata = pd.read_csv(file_path_metadata)
df_metadata = pd.DataFrame(df_metadata)
df_metadata

merged_df = df_clicks.merge(df_metadata, left_on='click_article_id', right_on='article_id', how='left')
merged_df

df_grouped = merged_df.groupby('user_id').agg({
    'click_article_id': list,
    'category_id': list
}).reset_index()
df_grouped

from collections import Counter

# Initialisation d'un dictionnaire pour stocker les représentations utilisateur
user_profiles = {}

# Parcourir le dataframe pour remplir les profils utilisateur
for index, row in df_grouped.iterrows():
    user_id = row['user_id']
    categories = row['category_id']

    if user_id not in user_profiles:
        user_profiles[user_id] = Counter()

    # Ajouter les catégories au profil de l'utilisateur
    user_profiles[user_id].update(categories)

# Convertir les profils utilisateur en dataframe
user_profiles_df = pd.DataFrame(user_profiles).fillna(0).T
user_profiles_df

pickle = pickle.reset_index()

df_temporary = df_metadata.copy()
df_temporary = df_temporary.drop(columns=["created_at_ts", "publisher_id", "words_count"])

pickle = pickle.merge(df_temporary, left_on='index', right_on='article_id', how='left')
pickle = pickle.drop(columns=["article_id"])
pickle = pickle.rename(columns={"index": "article_id"})
cols = list(pickle.columns)
cols.insert(1, cols.pop())
pickle = pickle[cols]

pickle

import pandas as pd
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity

# Exemple de DataFrame
dataX = {
    'article_id': [1, 2, 3, 4, 5],
    'embedding': [np.random.rand(250) for _ in range(5)],  # embeddings aléatoires pour l'exemple
    'label': ['A', 'B', 'C', 'A', 'B']
}
dfX = pd.DataFrame(dataX)
dfX


# Embeddings des articles que l'utilisateur aime (par exemple, articles 1 et 3)
liked_articles_embeddings = dfX[dfX['article_id'].isin([1, 3])]['embedding'].values
liked_articles_embeddingsX = pd.DataFrame(liked_articles_embeddings)
liked_articles_embeddingsX

# Calcul de l'embedding moyen de l'utilisateur
user_embedding = np.mean(liked_articles_embeddings, axis=0)
user_embedding

# Calcul de la similarité cosinus entre l'embedding de l'utilisateur et les embeddings des articles
dfX['similarity'] = dfX['embedding'].apply(lambda x: cosine_similarity([user_embedding], [x])[0][0])

# Tri des articles par similarité décroissante
recommended_articles = dfX.sort_values(by='similarity', ascending=False)

print(recommended_articles[['article_id', 'similarity']])
